requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
django-filter==24.5
python-decouple==3.8
psycopg2-binary==2.9.10
Pillow==10.4.0
python-dotenv==1.0.1
pandas==2.3.0 